Rename e2e to libdnstest, dummy to example
